perf(router): reduce the number of guard files
