--- src/router/guard/pageGuard.ts
@@ -1,18 +1,0 @@
-import type { Router } from 'vue-router';
-import { setRouteChange } from '/@/logics/mitt/routeChange';
-
-export function createPageGuard(router: Router) {
-  const loadedPageMap = new Map<string, boolean>();
-
-  router.beforeEach(async (to) => {
-    to.meta.loaded = !!loadedPageMap.get(to.path);
-    // Notify routing changes
-    setRouteChange(to);
-
-    return true;
-  });
-
-  router.afterEach((to) => {
-    loadedPageMap.set(to.path, true);
-  });
-}
