+++ src/store/index.ts
@@ -1,5 +1,6 @@
 import type { App } from 'vue';
 import { createPinia } from 'pinia';
+
 const store = createPinia();
 
 export function setupStore(app: App<Element>) {
